chore: rename contentLakeStore => localDataset
